+++ src/entitys/main/recommendedPlaylistImage.entity.ts
@@ -13,7 +13,7 @@ import { Exclude } from 'class-transformer';
 @Entity({ name: 'recommended_playlist_image' })
 export class RecommendedPlaylistImageEntity extends BaseEntity {
   @Exclude()
-  @ApiProperty({ type: 'bigint' })
+  @ApiProperty({ type: 'int' })
   @PrimaryGeneratedColumn({ type: 'bigint' })
   id: number;
