+++ src/entitys/main/recommendedPlaylistSongs.entity.ts
@@ -14,7 +14,7 @@ import { Exclude } from 'class-transformer';
 @Entity({ name: 'recommended_playlist_songs' })
 export class RecommendedPlaylistSongsEntity extends BaseEntity {
   @Exclude()
-  @ApiProperty({ type: 'bigint' })
+  @ApiProperty({ type: 'int' })
   @PrimaryGeneratedColumn({ type: 'bigint' })
   id: number;
 
@@ -34,7 +34,7 @@ export class RecommendedPlaylistSongsEntity extends BaseEntity {
   @JoinColumn({ name: 'song_id', referencedColumnName: 'id' })
   song: SongsEntity;
 
-  @ApiProperty({ type: 'bigint' })
+  @ApiProperty({ type: 'int' })
   @Column({ type: 'bigint' })
   order: number;
 }
